+++ pagespeed/envoy/envoy_url_async_fetcher.h
@@ -34,11 +34,8 @@
 #include "pagespeed/kernel/base/pool.h"
 #include "pagespeed/kernel/base/string.h"
 #include "pagespeed/kernel/base/thread_system.h"
-#include "pagespeed_remote_data_fetcher.h"
 #include "envoy_fetch.h"
-
-#include "external/envoy_api/envoy/api/v2/core/http_uri.pb.h"
-
+#include "envoy_cluster_manager.h"
 
 namespace net_instaweb {
 
@@ -70,15 +67,6 @@ struct EnvoyStats {
   // When we last checked the ultimate failure/success numbers for a
   // possible concern.
   static const char kEnvoyFetchLastCheckTimestampMs[];
-};
-
-class PagespeedDataFetcherCallback : public PagespeedRemoteDataFetcherCallback {
-public:
-  // Config::DataFetcher::RemoteDataFetcherCallback
-  void onSuccess(const std::string& data) override;
-
-  // Config::DataFetcher::RemoteDataFetcherCallback
-  void onFailure(FailureReason reason) override;
 };
 
 class EnvoyUrlAsyncFetcher : public UrlAsyncFetcher {
@@ -139,8 +127,8 @@ private:
 
   EnvoyFetchPool active_fetches_;
   // Add the pending task to this list
-  std::unique_ptr<PagespeedRemoteDataFetcher> PagespeedRemoteDataFetcherPtr;
-  EnvoyClusterManager* cluster_manager_;
+  //std::unique_ptr<EnvoyClusterManager> cluster_manager_;
+  EnvoyClusterManager *cluster_manager_;
   EnvoyFetchPool pending_fetches_;
   EnvoyFetchPool completed_fetches_;
   char* proxy_;
