+++ pagespeed/envoy/pagespeed_remote_data_fetcher.cc
@@ -30,14 +30,18 @@ void PagespeedRemoteDataFetcher::fetch() {
 }
 
 void PagespeedRemoteDataFetcher::onSuccess(Envoy::Http::MessagePtr&& response) {
-  callback_.onSuccess(response->body()->toString());
+  std::cout << "PagespeedRemoteDataFetcher::onSuccess data:" << response->body()->toString() << "\n";
+  std::cout.flush();
+  // callback_.onSuccess(response->body()->toString());
   request_ = nullptr;
 }
 
 void PagespeedRemoteDataFetcher::onFailure(Envoy::Http::AsyncClient::FailureReason reason) {
   // ENVOY_LOG(debug, "fetch remote data [uri = {}]: network error {}", uri_.uri(), enumToInt(reason));
+  std::cout << "PagespeedRemoteDataFetcher::onFailure \n";
+  std::cout.flush();
   request_ = nullptr;
-  callback_.onFailure(FailureReason::Network);
+  // callback_.onFailure(FailureReason::Network);
 }
 
 } // namespace net_instaweb
